==> research/02_data_validation.ipynb <==
import os

%pwd


os.chdir("../")


%pwd


import pandas as pd


file1_path = 'artifacts/data_ingestion/data/sales.csv'
df_sales = pd.read_csv(file1_path)
df_sales.drop(columns=["Unnamed: 0","transaction_id"], inplace=True, errors='ignore')
df_sales.info()

file2_path = 'artifacts/data_ingestion/data/sensor_stock_levels.csv'
df_sensor_stock_levels = pd.read_csv(file2_path)
df_sensor_stock_levels.drop(columns=["Unnamed: 0","id"], inplace=True, errors='ignore')
df_sensor_stock_levels.info()

file3_path = 'artifacts/data_ingestion/data/sensor_storage_temperature.csv'
df_sensor_storage_temperature = pd.read_csv(file3_path)
df_sensor_storage_temperature.drop(columns=["Unnamed: 0","id"], inplace=True, errors='ignore')
df_sensor_storage_temperature.info()

def process_and_save_csv(input_folder, output_folder):
    # Create the output folder if it doesn't exist
    os.makedirs(output_folder, exist_ok=True)

    # List all CSV files in the input folder
    csv_files = [file for file in os.listdir(input_folder) if file.endswith('.csv')]

    # Process each CSV file
    for csv_file in csv_files:
        # Construct the full path for input and output files
        input_path = os.path.join(input_folder, csv_file)
        output_path = os.path.join(output_folder, f"Dropped_validated_{csv_file}")

        # Load the CSV file into a DataFrame
        df = pd.read_csv(input_path)
        
        # Drop different columns in each CSV file
        if 'Unnamed: 0' in df.columns:
            df = df.drop(columns=['Unnamed: 0'])

        if 'id' in df.columns:
            df = df.drop(columns=['id'])

        if 'transaction_id' in df.columns:
            df = df.drop(columns=['transaction_id'])
            

        # Save the modified DataFrame to a new CSV file
        df.to_csv(output_path, index=False)

input_folder = 'artifacts/data_ingestion/data'
output_folder = 'artifacts/data_validation/validated_data_after_dropping_columns'

process_and_save_csv(input_folder, output_folder)


df1 = pd.read_csv('artifacts/data_validation/validated_data_after_dropping_columns/Dropped_validated_sales.csv')
df2 = pd.read_csv('artifacts/data_validation/validated_data_after_dropping_columns/Dropped_validated_sensor_stock_levels.csv')
df3 = pd.read_csv('artifacts/data_validation/validated_data_after_dropping_columns/Dropped_validated_sensor_storage_temperature.csv')

def convert_to_datetime(data: pd.DataFrame = None, column: str = None):

    dummy = data.copy()
    dummy[column] = pd.to_datetime(dummy[column], format='%Y-%m-%d %H:%M:%S')
    return dummy

sales_df = convert_to_datetime(df1, 'timestamp')
sales_df.info()

stock_df = convert_to_datetime(df2, 'timestamp')
stock_df.info()

temp_df = convert_to_datetime(df3, 'timestamp')
temp_df.info()

from datetime import datetime

def convert_timestamp_to_hourly(data: pd.DataFrame = None, column: str = None):
    dummy = data.copy()
    new_ts = dummy[column].tolist()
    new_ts = [i.strftime('%Y-%m-%d %H:00:00') for i in new_ts]
    new_ts = [datetime.strptime(i, '%Y-%m-%d %H:00:00') for i in new_ts]
    dummy[column] = new_ts
    return dummy

sales = convert_timestamp_to_hourly(sales_df, 'timestamp')
sales.head()

sales_agg = sales.groupby(['timestamp', 'product_id']).agg({'quantity': 'sum'}).reset_index()
sales_agg.head()

len(sales_agg)

stock = convert_timestamp_to_hourly(stock_df, 'timestamp')
stock.head()

stock_agg = stock.groupby(['timestamp', 'product_id']).agg({'estimated_stock_pct': 'mean'}).reset_index()
stock_agg.head()

len(stock_agg)

temp = convert_timestamp_to_hourly(temp_df, 'timestamp')
temp.head()

temp_agg = temp.groupby(['timestamp']).agg({'temperature': 'mean'}).reset_index()
temp_agg.head()

len(temp_agg)

merged_df = stock_agg.merge(sales_agg, on=['timestamp', 'product_id'], how='left')
merged_df.head()

merged_df = merged_df.merge(temp_agg, on='timestamp', how='left')
merged_df.head()

merged_df['quantity'] = merged_df['quantity'].fillna(0)
merged_df.info()

product_categories = sales_df[['product_id', 'category']]
product_categories = product_categories.drop_duplicates()

product_price = sales_df[['product_id', 'unit_price']]
product_price = product_price.drop_duplicates()

merged_df = merged_df.merge(product_categories, on="product_id", how="left")
merged_df.head()

merged_df = merged_df.merge(product_price, on="product_id", how="left")
merged_df.head()

merged_df['timestamp_day_of_month'] = merged_df['timestamp'].dt.day
merged_df['timestamp_day_of_week'] = merged_df['timestamp'].dt.dayofweek
merged_df['timestamp_hour'] = merged_df['timestamp'].dt.hour
merged_df.drop(columns=['timestamp'], inplace=True)
merged_df.head()

merged_df.drop(columns=['product_id'], inplace=True)
merged_df.head()

merged_df.isnull().sum()

import os
import pandas as pd
from datetime import datetime



def convert_to_datetime(data: pd.DataFrame = None, column: str = None):

    dummy = data.copy()
    dummy[column] = pd.to_datetime(dummy[column], format='%Y-%m-%d %H:%M:%S')
    return dummy


def convert_timestamp_to_hourly(data: pd.DataFrame = None, column: str = None):
    dummy = data.copy()
    new_ts = dummy[column].tolist()
    new_ts = [i.strftime('%Y-%m-%d %H:00:00') for i in new_ts]
    new_ts = [datetime.strptime(i, '%Y-%m-%d %H:00:00') for i in new_ts]
    dummy[column] = new_ts
    return dummy


def merge_and_save(input_file1, input_file2, input_file3, output_directory):
    # Read input files into dataframes
    df1 = pd.read_csv(input_file1)
    df2 = pd.read_csv(input_file2)
    df3 = pd.read_csv(input_file3)

    sales_df = convert_to_datetime(df1, 'timestamp')
    stock_df = convert_to_datetime(df2, 'timestamp')
    temp_df = convert_to_datetime(df3, 'timestamp')

    sales = convert_timestamp_to_hourly(sales_df, 'timestamp')
    sales_agg = sales.groupby(['timestamp', 'product_id']).agg({'quantity': 'sum'}).reset_index()

    stock = convert_timestamp_to_hourly(stock_df, 'timestamp')
    stock_agg = stock.groupby(['timestamp', 'product_id']).agg({'estimated_stock_pct': 'mean'}).reset_index()

    temp = convert_timestamp_to_hourly(temp_df, 'timestamp')
    temp_agg = temp.groupby(['timestamp']).agg({'temperature': 'mean'}).reset_index()

    merged_df = stock_agg.merge(sales_agg, on=['timestamp', 'product_id'], how='left')
    merged_df = merged_df.merge(temp_agg, on='timestamp', how='left')

    merged_df['quantity'] = merged_df['quantity'].fillna(0)

    product_categories = sales_df[['product_id', 'category']]
    product_categories = product_categories.drop_duplicates()

    product_price = sales_df[['product_id', 'unit_price']]
    product_price = product_price.drop_duplicates()

    merged_df = merged_df.merge(product_categories, on="product_id", how="left")
    merged_df = merged_df.merge(product_price, on="product_id", how="left")

    merged_df['timestamp_day_of_month'] = merged_df['timestamp'].dt.day
    merged_df['timestamp_day_of_week'] = merged_df['timestamp'].dt.dayofweek
    merged_df['timestamp_hour'] = merged_df['timestamp'].dt.hour
    merged_df.drop(columns=['timestamp','product_id'], inplace=True)



    # Create the output directory if it doesn't exist
    os.makedirs(output_directory, exist_ok=True)

    # Define the output file path
    output_file_path = os.path.join(output_directory, 'merged_output.csv')

    # Save the merged dataframe to the specified directory
    merged_df.to_csv(output_file_path, index=False)

    print(f'Merged dataframe saved to: {output_file_path}')

# Example usage:
merge_and_save('artifacts/data_validation/validated_data_after_dropping_columns/Dropped_validated_sales.csv', 
               'artifacts/data_validation/validated_data_after_dropping_columns/Dropped_validated_sensor_stock_levels.csv', 
               'artifacts/data_validation/validated_data_after_dropping_columns/Dropped_validated_sensor_storage_temperature.csv', 
               'artifacts/data_validation/validated_data_after_dropping_columns')



df_dtypes = pd.read_csv('artifacts/data_validation/validated_data_after_dropping_columns/merged_output.csv')

df_dtypes.dtypes

%pwd

from dataclasses import dataclass
from pathlib import Path


@dataclass(frozen=True)
class DataValidationConfig:
    root_dir: Path
    STATUS_FILE: str
    all_schema: dict
    input_folder : Path
    output_folder : Path
    input_file1 : Path
    input_file2 : Path
    input_file3 : Path
    output_directory : Path
    merged_output: Path

from src.mlProject.constants import *
from src.mlProject.utils.common import read_yaml, create_directories,convert_to_datetime,convert_timestamp_to_hourly

class ConfigurationManager:
    def __init__(
        self,
        config_filepath = CONFIG_FILE_PATH,
        params_filepath = PARAMS_FILE_PATH,
        schema_filepath = SCHEMA_FILE_PATH):

        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])

    
    def get_data_validation_config(self) -> DataValidationConfig:
        config = self.config.data_validation
        schema = self.schema.COLUMNS

        create_directories([config.root_dir])

        data_validation_config = DataValidationConfig(
            root_dir=config.root_dir,
            input_folder = config.input_folder,
            output_folder = config.output_folder,
            input_file1 = config.input_file1,
            input_file2 = config.input_file2,
            input_file3 = config.input_file3,
            output_directory = config.output_directory,
            merged_output = config.merged_output,
            STATUS_FILE = config.STATUS_FILE,
            all_schema=schema
            )

        return data_validation_config

import os
from mlProject import logger
import pandas as pd
from sklearn.model_selection import train_test_split


class DataValiadtion:
    def __init__(self, config: DataValidationConfig):
        self.config = config

    def process_and_save_csv(self):

        try:

            # Create the output folder if it doesn't exist
            os.makedirs(self.config.output_folder, exist_ok=True)

            # List all CSV files in the input folder
            csv_files = [file for file in os.listdir(self.config.input_folder) if file.endswith('.csv')]

            # Process each CSV file
            for csv_file in csv_files:
                # Construct the full path for input and output files
                input_path = os.path.join(self.config.input_folder, csv_file)
                output_path = os.path.join(self.config.output_folder, f"Dropped_validated_{csv_file}")

                # Load the CSV file into a DataFrame
                df = pd.read_csv(input_path)
                
                # Drop different columns in each CSV file
                if 'Unnamed: 0' in df.columns:
                    df = df.drop(columns=['Unnamed: 0'])

                if 'id' in df.columns:
                    df = df.drop(columns=['id'])

                if 'transaction_id' in df.columns:
                    df = df.drop(columns=['transaction_id'])

            # Save the modified DataFrame to a new CSV file
            df.to_csv(output_path, index=False)
        
        except Exception as e:
            raise e
        


    def merge_and_save(self):
        try:
            # Read input files into dataframes
            df1 = pd.read_csv(self.config.input_file1)
            df2 = pd.read_csv(self.config.input_file2)
            df3 = pd.read_csv(self.config.input_file3)

            sales_df = convert_to_datetime(df1, 'timestamp')
            stock_df = convert_to_datetime(df2, 'timestamp')
            temp_df = convert_to_datetime(df3, 'timestamp')

            sales = convert_timestamp_to_hourly(sales_df, 'timestamp')
            sales_agg = sales.groupby(['timestamp', 'product_id']).agg({'quantity': 'sum'}).reset_index()

            stock = convert_timestamp_to_hourly(stock_df, 'timestamp')
            stock_agg = stock.groupby(['timestamp', 'product_id']).agg({'estimated_stock_pct': 'mean'}).reset_index()

            temp = convert_timestamp_to_hourly(temp_df, 'timestamp')
            temp_agg = temp.groupby(['timestamp']).agg({'temperature': 'mean'}).reset_index()

            merged_df = stock_agg.merge(sales_agg, on=['timestamp', 'product_id'], how='left')
            merged_df = merged_df.merge(temp_agg, on='timestamp', how='left')

            merged_df['quantity'] = merged_df['quantity'].fillna(0)

            product_categories = sales_df[['product_id', 'category']]
            product_categories = product_categories.drop_duplicates()

            product_price = sales_df[['product_id', 'unit_price']]
            product_price = product_price.drop_duplicates()

            merged_df = merged_df.merge(product_categories, on="product_id", how="left")
            merged_df = merged_df.merge(product_price, on="product_id", how="left")

            merged_df['timestamp_day_of_month'] = merged_df['timestamp'].dt.day
            merged_df['timestamp_day_of_week'] = merged_df['timestamp'].dt.dayofweek
            merged_df['timestamp_hour'] = merged_df['timestamp'].dt.hour
            merged_df.drop(columns=['timestamp','product_id'], inplace=True)


            # Create the output directory if it doesn't exist
            os.makedirs(self.config.output_directory, exist_ok=True)

            # Define the output file path
            output_file_path = os.path.join(self.config.output_directory, 'merged_output.csv')

            # Save the merged dataframe to the specified directory
            merged_df.to_csv(output_file_path, index=False)

        except Exception as e:
            raise e



    def validate_all_columns(self)-> bool:
        try:
            validation_status = None

            data = pd.read_csv(self.config.merged_output)
            all_cols = list(data.columns)

            all_schema = self.config.all_schema.keys()

            
            for col in all_cols:
                if col not in all_schema:
                    validation_status = False
                    with open(self.config.STATUS_FILE, 'w') as f:
                        f.write(f"Validation status: {validation_status}")
                else:
                    validation_status = True
                    with open(self.config.STATUS_FILE, 'w') as f:
                        f.write(f"Validation status: {validation_status}")

            return validation_status
        
        except Exception as e:
            raise e
        
    def train_test_spliting(self):
        data = pd.read_csv(self.config.merged_output)

        # Split the data into training and test sets. (0.75, 0.25) split.
        train, test = train_test_split(data)

        train.to_csv(os.path.join(self.config.root_dir, "train.csv"),index = False)
        test.to_csv(os.path.join(self.config.root_dir, "test.csv"),index = False)

        logger.info("Splited data into training and test sets")
        logger.info(train.shape)
        logger.info(test.shape)


try:
    config = ConfigurationManager()
    data_validation_config = config.get_data_validation_config()
    data_validation = DataValiadtion(config=data_validation_config)
    data_validation.process_and_save_csv()
    data_validation.merge_and_save()
    data_validation.validate_all_columns()
    data_validation.train_test_spliting()

    
except Exception as e:
    raise e
