==> research/03_data_transformation.ipynb <==
import os

%pwd


os.chdir("../")


%pwd


from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline
from sklearn.impute import SimpleImputer
from sklearn.preprocessing import StandardScaler,OrdinalEncoder,OneHotEncoder


# Categorigal Pipeline
categorical_cols_for_OrdinalEncoding = ['category']

Numerical_col_1  = ['temperature','unit_price', 'timestamp_day_of_month', 'timestamp_day_of_week','timestamp_hour']

Numerical_col_2  = ['quantity']

categorical_col_for_OnehotEncoding = ['category']


Numerical_pipeline_one_for_missingvalues = Pipeline(
                steps=[
                ('scaler',StandardScaler())
                 ])

Numerical_pipeline_two_for_missingvalues = Pipeline(
                steps=[
                ('imputer',SimpleImputer(strategy='constant', fill_value=0)),
                ('scaler',StandardScaler())
                 ])
            
categorical_pipeline_for_OnehotEncoding = Pipeline(
                            steps=[
                            ('one_hot_encoder', OneHotEncoder(sparse_output=False,handle_unknown = 'ignore')),
                            ('scaler',StandardScaler())
                            ]
                            )

preprocessor=ColumnTransformer(transformers= 
        [('Numerical_pipeline_1_for_missingvalues', Numerical_pipeline_one_for_missingvalues,Numerical_col_1),
         ('Numerical_pipeline_2_for_missingvalues', Numerical_pipeline_two_for_missingvalues,Numerical_col_2),
         ('categorical_pipeline_for_OnehotEncoding', categorical_pipeline_for_OnehotEncoding,categorical_col_for_OnehotEncoding )
         ],
                                                          remainder='passthrough',sparse_threshold=0)     





import pandas as pd

merged_df = pd.read_csv('artifacts/data_validation/validated_data_after_dropping_columns/merged_output.csv')

merged_df.head()

preprocessor.fit_transform(merged_df)


df_final = pd.DataFrame(preprocessor.fit_transform(merged_df))


df_final.head()

df_final.shape

from sklearn.model_selection import train_test_split

train, test = train_test_split(df_final)

train.shape

test.shape

import os
from src.mlProject import logger
from sklearn.model_selection import train_test_split
import pandas as pd
import numpy as np


from src.mlProject.constants import *
from src.mlProject.utils.common import *

from dataclasses import dataclass
from pathlib import Path

@dataclass(frozen=True)
class DataTransformationConfig:
    root_dir: Path
    train_file_path: Path
    test_file_path: Path
    transformed_object_file_path: Path
    transformed_train_file_path: Path
    transformed_test_file_path: Path

class ConfigurationManager:
    def __init__(
        self,
        config_filepath = CONFIG_FILE_PATH,
        params_filepath = PARAMS_FILE_PATH,
        schema_filepath = SCHEMA_FILE_PATH):

        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])

    
    def get_data_transformation_config(self) -> DataTransformationConfig:
        config = self.config.data_transformation

        create_directories([config.root_dir])

        data_transformation_config = DataTransformationConfig(
            root_dir=config.root_dir,
            train_file_path = config.train_file_path,
            test_file_path = config.test_file_path,
            transformed_object_file_path = config.transformed_object_file_path,
            transformed_train_file_path = config.transformed_train_file_path,
            transformed_test_file_path = config.transformed_test_file_path
        )

        return data_transformation_config

from src.mlProject.entity.config_entity import DataTransformationConfig
import os
from src.mlProject import logger
from sklearn.model_selection import train_test_split
import pandas as pd


class DataTransformation:
    def __init__(self, config: DataTransformationConfig):
        self.config = config



    def get_data_transformer_object(self):
        try:

            Numerical_col_1  = ['temperature','unit_price', 'timestamp_day_of_month', 'timestamp_day_of_week','timestamp_hour']

            Numerical_col_2  = ['quantity']

            categorical_col_for_OnehotEncoding = ['category']


            Numerical_pipeline_one_for_missingvalues = Pipeline(
                            steps=[
                            ('scaler',StandardScaler())
                            ])

            Numerical_pipeline_two_for_missingvalues = Pipeline(
                            steps=[
                            ('imputer',SimpleImputer(strategy='constant', fill_value=0)),
                            ('scaler',StandardScaler())
                            ])
                        
            categorical_pipeline_for_OnehotEncoding = Pipeline(
                                        steps=[
                                        ('one_hot_encoder', OneHotEncoder(sparse_output=False,handle_unknown = 'ignore')),
                                        ('scaler',StandardScaler())
                                        ]
                                        )

            preprocessor=ColumnTransformer(transformers= 
                    [('Numerical_pipeline_1_for_missingvalues', Numerical_pipeline_one_for_missingvalues,Numerical_col_1),
                    ('Numerical_pipeline_2_for_missingvalues', Numerical_pipeline_two_for_missingvalues,Numerical_col_2),
                    ('categorical_pipeline_for_OnehotEncoding', categorical_pipeline_for_OnehotEncoding,categorical_col_for_OnehotEncoding )
                    ],
                                                                    remainder='passthrough',sparse_threshold=0)     
            
            return preprocessor
            

        except Exception as e:
            raise e
    
    def initiate_data_transformation(self,):
        try:
            
            train_df = pd.read_csv(self.config.train_file_path)
            test_df = pd.read_csv(self.config.test_file_path)
            preprocessor = self.get_data_transformer_object()


            #training dataframe
            input_feature_train_df = train_df.drop(columns='estimated_stock_pct', axis=1)
            target_feature_train_df = train_df['estimated_stock_pct']

            #testing dataframe
            input_feature_test_df = test_df.drop(columns='estimated_stock_pct', axis=1)
            target_feature_test_df = test_df['estimated_stock_pct']

            preprocessor_object = preprocessor.fit(input_feature_train_df)
            input_feature_train_final = preprocessor_object.transform(input_feature_train_df)
            input_feature_test_final =preprocessor_object.transform(input_feature_test_df)


            train_arr = np.c_[input_feature_train_final, np.array(target_feature_train_df)]
            test_arr = np.c_[input_feature_test_final, np.array(target_feature_test_df)]


            #save numpy array data

            save_numpy_array(train_arr, self.config.transformed_train_file_path, 'train_array.npy')
            save_numpy_array(test_arr, self.config.transformed_train_file_path, 'test_array.npy')
            save_preprocessor(preprocessor_object, self.config.transformed_object_file_path)

            
            return (
                train_arr,
                test_arr,
                self.config.transformed_object_file_path, preprocessor_object
            )
        
        except Exception as e:
            raise e
            



try:
    config = ConfigurationManager()
    data_transformation_config = config.get_data_transformation_config()
    data_transformation = DataTransformation(config=data_transformation_config)
    data_transformation.get_data_transformer_object()
    data_transformation.initiate_data_transformation()

except Exception as e:
    raise e
